pyproject.toml
[project]
name = "neocarta"
version = "0.1.0"
description = "Library that facilitates the creation of semantic layer metadata graph in Neo4j."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Alex Gilmore" }
]
requires-python = ">=3.10"
dependencies = [
    "google-cloud-bigquery[pandas]>=3.40.0",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "neo4j>=6.1.0",
    "pandas>=3.0.0",
    "google-cloud-dataplex>=2.16.0",
    "sqlglot>=28.10.1",
    "google-cloud-bigquery-storage>=2.36.2",
]

[dependency-groups]
# Development tools
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
    "testcontainers[neo4j]>=4.14.2",
    "ruff>=0.15.10",
]

# Local MCP server for SQL metadata retrieval
mcp-server = [
    "mcp-server",
]

# Text2SQL agent with LangChain (includes mcp-server dependencies)
agent = [
    "mcp-server",
    "google-cloud-bigquery>=3.40.0",
    "langchain>=1.2.7",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
]
eval = [
    "fastmcp>=2.14.4",
    "mcp>=1.26.0",
    "tiktoken>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.workspace]
members = ["mcp_server"]

[tool.uv.sources]
mcp-server = { workspace = true }

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "C90",     # McCabe complexity
    "COM812",  # Conflicts with the formatter
    "CPY",     # No copyright
    "D205",    # Blank line after summary — multiline summaries add clarity
    "D212",    # Conflicts with Google convention (D213)
    "FIX002",  # Line contains TODO
    "PERF203", # Rarely useful
    "PLR09",   # Too many something (args, statements, etc.)
    "TD002",   # Missing author in TODO
    "TD003",   # Missing issue link in TODO

    # Intentionally lenient
    "ANN401",  # Any in type annotations
    "ARG001",  # Unused function arguments — needed for interface conformance
    "BLE",     # Blind exceptions
    "B006",    # Mutable default arguments — lists are read-only in this codebase
    "E501",    # Line too long — unavoidable in data/SQL code
    "INP001",  # Implicit namespace packages — intentional
    "PLR2004", # Magic value comparisons — too noisy in data code
    "SLF001",  # Private member access — often necessary
    "T201",    # print — will migrate to logging in a future PR
    "TID252",  # Relative imports are fine

    # Boolean handling
    "FBT001",  # Boolean type hint positional argument
    "FBT002",  # Boolean default value positional argument
    "FBT003",  # Boolean positional value in call

    # Stylistic pandas preferences — existing style is fine
    "PD003",   # .isnull() vs .isna()
    "PD011",   # .values vs .to_numpy()
    "PERF401", # Replace for loop with list comprehension — often less readable

    # Internal SQL construction, not user-facing injection risk
    "S608",    # Hardcoded SQL expression
    # Intentional silent pass (e.g. best-effort parsing)
    "S110",    # try-except-pass
    "SIM117",  # Merge nested with statements — nested is often more readable

    # Exception details
    "EM101",   # Raw string in exception
    "EM102",   # f-string in exception
    "TRY003",  # Long exception messages outside exception class
    "TRY300",  # Consider moving return to else block — reduces readability
    "TRY301",  # Abstract raise to inner function — overly pedantic
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # assert is idiomatic in pytest
    "ANN",     # Type hints optional in tests
    "D",       # Docstrings optional in tests
    "N999",    # _IT suffix is a deliberate integration-test naming convention
    "PLC0415", # Lazy imports acceptable in tests
]
"eval/retrievers/bigquery_schema_retriever.py" = [
    "PLC0415", # Lazy import to avoid circular dependency with eval package
]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["field_validator", "model_validator"]

[tool.ruff.lint.pydocstyle]
convention = "google"
ignore-var-parameters = true  # ignore missing documentation for *args and **kwargs parameters
